refactor: Make UndoRedoStack non-SO and move to DungeonCrawlerData
