--- Unity Project/Dungeoneering/Assets/_Libs/DungeonCrawler.Unity.Editor/src/UndoRedoStackDataExtensions.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 783e3892cb5e17b44a3a4b5c1bab3099
