--- Unity Project/Dungeoneering/Assets/_Project/Data/Editor/MainUndoRedoStack.asset
@@ -1,14 +1,0 @@
-%YAML 1.1
-%TAG !u! tag:unity3d.com,2011:
---- !u!114 &11400000
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 0}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 2e83c0fccbffad248893d8a3bbc74e4c, type: 3}
-  m_Name: MainUndoRedoStack
-  m_EditorClassIdentifier:
