--- Unity Project/Dungeoneering/Assets/_Project/Data/Editor/MainUndoRedoStack.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: ebba17630ec3c1b4583c3d3527a5b966
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 11400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
