--- Unity Project/Dungeoneering/Assets/_Tests/Test Data/Test Undo Redo Stack.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: e1b958c313a4fd74cb57ee22dc184d49
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 11400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
